pyproject.toml
[project]
name = "whisper-hacking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "tts>=0.22.0",
    "torch==2.7.0+xpu; sys_platform != 'win32'",
    "torchaudio==2.7.0+xpu; sys_platform != 'win32'",
    "intel-extension-for-pytorch==2.7.10+xpu; sys_platform != 'win32'",
    "oneccl_bind_pt==2.7.0+xpu; sys_platform != 'win32'",
    "pre-commit>=4.2.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "black>=25.1.0",
    "mypy>=1.15.0",
]

[tool.uv]
index-url = "https://download.pytorch.org/whl/xpu"
extra-index-url = [
    "https://pytorch-extension.intel.com/release-whl/stable/xpu/us/",
    "https://pypi.org/simple"
]
index-strategy = "unsafe-best-match"
